7-Zip Version 4.25 Beta
-  More 64-bit compatibilty

+++ 7zip/Archive/7z/7zEncode.cpp
@@ -258,7 +258,7 @@ HRESULT CEncoder::CreateMixerCoder()
       outStreamSpec->Init();
       writeCoderProperties->WriteCoderProperties(outStream);
       
-      UInt32 size = outStreamSpec->GetSize();
+      size_t size = outStreamSpec->GetSize();
       
       // encodingInfo.Properties.SetCapacity(size);
       if (encodingInfo.AltCoders.Size() == 0)
